+++ R/theme_blackboard.R
@@ -3,6 +3,7 @@
 #' A modern, sleek and dark theme for ggplot.
 #'
 #' @inheritParams theme_modern
+#' @inherit theme_modern note
 #'
 #' @examples
 #' library(ggplot2)
@@ -12,23 +13,23 @@
 #'   geom_point(color = "white") +
 #'   theme_blackboard()
 #' @export
-theme_blackboard <-
-  function(base_size = 11,
-           base_family = "",
-           plot.title.size = 15,
-           plot.title.face = "plain",
-           plot.title.space = 20,
-           plot.title.position = "plot",
-           legend.position = "right",
-           axis.title.space = 20,
-           legend.title.size = 13,
-           legend.text.size = 12,
-           axis.title.size = 13,
-           axis.title.face = "plain",
-           axis.text.size = 12,
-           axis.text.angle = NULL,
-           tags.size = 15,
-           tags.face = "bold") {
+theme_blackboard <- function(base_size = 11,
+                             base_family = "",
+                             plot.title.size = 1.35 * base_size,
+                             plot.title.face = "plain",
+                             plot.title.space = 1.8 * base_size,
+                             plot.title.position = "plot",
+                             legend.position = "right",
+                             axis.title.space = 1.8 * base_size,
+                             legend.title.size = 1.2 * base_size,
+                             legend.text.size = 1.1 * base_size,
+                             axis.title.size = 1.2 * base_size,
+                             axis.title.face = "plain",
+                             axis.text.size = 1.1 * base_size,
+                             axis.text.angle = NULL,
+                             tags.size = 1.35 * base_size,
+                             tags.face = "bold",
+                            ...) {
     theme_modern(
       base_size = base_size,
       base_family = base_family,
@@ -51,9 +52,11 @@ theme_blackboard <-
         plot.background = element_rect(fill = "#0d0d0d"),
         panel.background = element_rect(fill = "#0d0d0d"),
         legend.background = element_rect(fill = "#0d0d0d"),
-        axis.line = element_line(color = "#E0E0E0"),
+        axis.line = element_line(color = "#E0E0E0", linewidth = base_size / 22),
+        rect = element_rect(colour = "#E0E0E0", linewidth = base_size / 22),
         text = element_text(color = "#E0E0E0"),
         axis.text = element_text(color = "#E0E0E0"),
-        strip.text = element_text(color = "#E0E0E0")
+        strip.text = element_text(color = "#E0E0E0"),
+        ...
       )
   }